nvim-cmp mappings
Remove tab and enter from my list of mappings. They don't help and often
get in the way. Just going to stick with <c-n> and <c-p> to cycle
through the list and <c-y> to select an option. This way there's no
chance of accidentally selecting an auto-complete option when just
trying to write a new line