Tmux stuff
Moved the leader key to Ctrl + T as Ctrl + J was a bit of a faff to hit.
Ctrl + T doesn't keybind to anything I care about in bash, neovim or
Windows Terminal

I also removed the code to 1 base window numbers. Since I use a custom
keyboard which does have 0 first, this was not a necessary optimisation
for me